restructured raw data processing dir
